+++ lib/Utilities_and_helper_widgets/translations_service.dart
@@ -2,9 +2,9 @@
 //*How to use this:
 ///1-import get in to your pubspec.yaml
 /*2-GetMaterialApp(
-        
+  
         title: "Akartkom",
-        home: SplashScreen1Sub(),
+        home: SplashScreenWidget(),
         translations: TranslationsService(), /// your translations
         locale:Locale('en', 'US'), /// translations will be displayed in that locale
         fallbackLocale: Locale('en', 'US'),
@@ -20,14 +20,12 @@
 
 import 'package:get/get.dart';
 
-
 class TranslationsService extends Translations {
   @override
   Map<String, Map<String, String>> get keys => {
         'en_US': {
           /// add your locales like this
         },
-        'ar_EG': {
-        }
+        'ar_EG': {}
       };
 }