+++ lib/widgets/text.dart
@@ -3,12 +3,12 @@ import 'package:flutter/material.dart';
 
 class Txt extends StatelessWidget {
   final String text;
-  final Color color;
+  final Color? color;
   final FontWeight weight;
   final double size;
   final TextAlign textAlign;
   const Txt(this.text,
-      {this.color = Colors.black,
+      {this.color ,
       this.weight = FontWeight.normal,
       this.size = 16,
       this.textAlign = TextAlign.center});